requirements.txt
aiogram==3.4.1
aiohttp==3.9.5
httpx==0.27.0
python-dotenv==1.0.1
